requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
groq==0.4.1
python-docx==1.1.0
PyPDF2==3.0.1